+++ R/replace.R
@@ -16,6 +16,25 @@ setMethod(
   }
 )
 
+#' @export
+#' @rdname missing
+#' @aliases replace_NA,data.frame-method
+setMethod(
+  f = "replace_NA",
+  signature = signature(x = "data.frame"),
+  definition = function(x, value = 0) {
+    x[] <- lapply(
+      X = x,
+      FUN = function(x, value) {
+        x[is.na(x)] <- value
+        x
+      },
+      value = value
+    )
+    x
+  }
+)
+
 ## Infinite values -------------------------------------------------------------
 #' @export
 #' @rdname infinite
@@ -29,6 +48,25 @@ setMethod(
   }
 )
 
+#' @export
+#' @rdname infinite
+#' @aliases replace_Inf,data.frame-method
+setMethod(
+  f = "replace_Inf",
+  signature = signature(x = "data.frame"),
+  definition = function(x, value = 0) {
+    x[] <- lapply(
+      X = x,
+      FUN = function(x, value) {
+        x[is.infinite(x)] <- value
+        x
+      },
+      value = value
+    )
+    x
+  }
+)
+
 ## Zeros -----------------------------------------------------------------------
 #' @export
 #' @rdname zero
@@ -38,6 +76,63 @@ setMethod(
   signature = signature(x = "matrix"),
   definition = function(x, value) {
     x[is_zero(x)] <- value
+    x
+  }
+)
+
+#' @export
+#' @rdname zero
+#' @aliases replace_zero,data.frame-method
+setMethod(
+  f = "replace_zero",
+  signature = signature(x = "data.frame"),
+  definition = function(x, value) {
+    num <- vapply(X = x, FUN = is.numeric, FUN.VALUE = logical(1))
+    nozero <- vapply(
+      X = x[, num],
+      FUN = function(x, value) {
+        x[is_zero(x)] <- value
+        x
+      },
+      FUN.VALUE = numeric(nrow(x)),
+      value = value
+    )
+    x[, num] <- nozero
+    x
+  }
+)
+
+## Empty string ----------------------------------------------------------------
+#' @export
+#' @rdname empty
+#' @aliases replace_empty,matrix-method
+setMethod(
+  f = "replace_empty",
+  signature = signature(x = "matrix"),
+  definition = function(x, value) {
+    x[is_empty_string(x)] <- value
+    x
+  }
+)
+
+#' @export
+#' @rdname empty
+#' @aliases replace_empty,data.frame-method
+setMethod(
+  f = "replace_empty",
+  signature = signature(x = "data.frame"),
+  definition = function(x, value) {
+    char <- vapply(X = x, FUN = is.character, FUN.VALUE = logical(1))
+    noblank <- vapply(
+      X = x[, char],
+      FUN = function(x, value) {
+        x[is_empty_string(x)] <- value
+        x
+      },
+      FUN.VALUE = character(nrow(x)),
+      value = value
+    )
+    x[, char] <- noblank
     x
   }
 )