+++ inst/examples/ex-empty.R
@@ -1,0 +1,13 @@
+## Create a data matrix
+X <- matrix(sample(LETTERS, 25, TRUE), nrow = 5, ncol = 5)
+
+## Add empty string
+k <- sample(1:25, 3, FALSE)
+X[k] <- ""
+X
+
+## Remove rows with empty strings
+remove_empty(X, margin = 1)
+
+## Replace empty strings
+replace_empty(X, value = "XXX")