+++ NEWS.md
@@ -1,6 +1,8 @@
 # arkhe 1.2.0.9000
 ## New classes and methods
 * Add `remove_constant()` to remove constant columns.
+* Add `remove_empty()` to remove empty strings in a `matrix`-like object.
+* Add `replace_empty()` to replace empty strings in a `matrix`-like object.
 
 ## Bugfixes & changes
 * Remove deprecated methods.